# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkyBooker API",
        "description": "Flight Booking System API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/v1/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "List all bookings with pagination",
                "operationId": "2f7e7260c708227e4c44de9d5b59ea36",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings with pagination"
                    }
                }
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "operationId": "9ec8aa3de0d1efdd0cfe72c27eac745c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "flight_id",
                                    "number_of_seats",
                                    "total_price",
                                    "status",
                                    "passengers"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "flight_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "number_of_seats": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 999.98
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Window seats preferred",
                                        "nullable": true
                                    },
                                    "special_requests": {
                                        "type": "string",
                                        "example": "Wheelchair assistance",
                                        "nullable": true
                                    },
                                    "passengers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+1234567890",
                                                    "nullable": true
                                                },
                                                "date_of_birth": {
                                                    "type": "string",
                                                    "example": "1990-01-01",
                                                    "nullable": true
                                                },
                                                "nationality": {
                                                    "type": "string",
                                                    "example": "USA"
                                                },
                                                "passport_number": {
                                                    "type": "string",
                                                    "example": "AB123456"
                                                },
                                                "id_number": {
                                                    "type": "string",
                                                    "example": "ID123456"
                                                },
                                                "seat_number": {
                                                    "type": "string",
                                                    "example": "12A",
                                                    "nullable": true
                                                },
                                                "meal_preference": {
                                                    "type": "string",
                                                    "example": "vegetarian",
                                                    "nullable": true
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/bookings/{booking}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get booking by ID",
                "operationId": "274c0a8ca69c9b806bbca9bb67511b30",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking retrieved successfully"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Update a booking",
                "operationId": "28ce32832486f9a9711254c32ea84dac",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "flight_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "number_of_seats": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1499.97
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "confirmed"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Updated notes"
                                    },
                                    "special_requests": {
                                        "type": "string",
                                        "example": "Updated requests"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Delete a booking",
                "operationId": "81743b411a3b756bb7340296d42b3317",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking deleted successfully"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            }
        },
        "/api/v1/bookings/{booking}/passengers": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get all passengers for a booking",
                "operationId": "3725b9308a913ada1b0f8b6bf54af1c0",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of passengers for the booking"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            }
        },
        "/api/v1/bookings/{booking}/tickets": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get all tickets for a booking",
                "operationId": "3cd1b5aa10493b6b57f7836ac3d8b603",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tickets for the booking"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            }
        },
...
    },
...
}